Extract OnChangeLambda to own file.
