+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/OpenAPIYamlGenerator.java
@@ -18,17 +18,15 @@
 package org.openapitools.codegen.languages;
 
 import org.openapitools.codegen.*;
+import org.openapitools.codegen.templating.mustache.OnChangeLambda;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.ImmutableMap.Builder;
-import com.samskivert.mustache.Mustache;
 import com.samskivert.mustache.Mustache.Lambda;
-import com.samskivert.mustache.Template;
+
 import io.swagger.v3.oas.models.Operation;
 
-import java.io.IOException;
-import java.io.Writer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -107,26 +105,6 @@ public class OpenAPIYamlGenerator extends DefaultCodegen implements CodegenConfi
     public String escapeUnsafeCharacters(String input) {
         // just return the original string
         return input;
-    }
-
-    /**
-     * Lambda writes current fragment to the output when it is different than
-     * previous fragment.
-     */
-    public static class OnChangeLambda implements Mustache.Lambda {
-        private static final Logger LOGGER = LoggerFactory.getLogger(OnChangeLambda.class);
-
-        private String lastVal = null;
-
-        @Override
-        public void execute(Template.Fragment frag, Writer out) throws IOException {
-            String curVal = frag.execute();
-            LOGGER.debug("[lastVal={}, curVal={}]", lastVal, curVal);
-            if (curVal != null && !curVal.equals(lastVal)) {
-                out.write(curVal);
-                lastVal = curVal;
-            }
-        }
     }
 
 }
